+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ v3.0.0
 ------
  - New [Default] Wallet Version 3  
    Better encryption scheme / key derivation.
+ - Deprecated passing in `primaryPrivateKey`, should use `primarySeed` instead.
 
 v2.0.1
 ------